+++ src/installer/mod.rs
@@ -1,3 +1,4 @@
+pub mod cpp;
 pub mod driver;
 pub mod generic;
 pub mod manager;